+++ ffi/tests/ffi_client_tests.rs
@@ -3,6 +3,7 @@ use glide_core::errors::RequestErrorType;
 use glide_core::request_type::RequestType;
 use glide_ffi::*;
 use protobuf::Message;
+use rstest::rstest;
 use std::collections::HashMap;
 use std::ffi::{c_char, c_ulong, c_void, CStr};
 use std::net::TcpListener;
@@ -15,22 +16,39 @@ use std::sync::{LazyLock, Mutex};
 use tokio::runtime::Runtime;
 use tokio::time::{sleep, Duration};
 
-static SUCCESS_COUNTER: LazyLock<Arc<AtomicUsize>> =
+static ASYNC_SUCCESS_COUNTER: LazyLock<Arc<AtomicUsize>> =
     LazyLock::new(|| Arc::new(AtomicUsize::new(0)));
-static FAILURE_COUNTER: LazyLock<Arc<AtomicUsize>> =
+static ASYNC_FAILURE_COUNTER: LazyLock<Arc<AtomicUsize>> =
     LazyLock::new(|| Arc::new(AtomicUsize::new(0)));
-type StringResultMap =
-    LazyLock<Mutex<HashMap<usize, Result<String, (&'static str, RequestErrorType)>>>>;
-static RESULTS_MAP: StringResultMap = LazyLock::new(|| Mutex::new(HashMap::new()));
+type StringResultMap = LazyLock<Mutex<HashMap<usize, Result<String, (String, RequestErrorType)>>>>;
+static ASYNC_RESULTS_MAP: StringResultMap = LazyLock::new(|| Mutex::new(HashMap::new()));
 
-/// Success callback function for String responses
+/// Success callback function for String responses for the async client
 extern "C" fn string_success_callback(index: usize, response_ptr: *const CommandResponse) {
+    let mut map = ASYNC_RESULTS_MAP
+        .lock()
+        .expect("Failed to aquire the results' lock");
+    map.insert(index, Ok(parse_string_res(response_ptr)));
+    ASYNC_SUCCESS_COUNTER.fetch_add(1, Ordering::SeqCst);
+}
+
+/// Failure callback function for the async client
+extern "C" fn failure_callback(
+    index: usize,
+    err_msg_ptr: *const c_char,
+    error_type: RequestErrorType,
+) {
+    let mut map = ASYNC_RESULTS_MAP
+        .lock()
+        .expect("Failed to aquire the results' lock");
+    map.insert(index, Err((parse_error_msg(err_msg_ptr), error_type)));
+    ASYNC_FAILURE_COUNTER.fetch_add(1, Ordering::SeqCst);
+}
+
+fn parse_string_res(response_ptr: *const CommandResponse) -> String {
     assert!(!response_ptr.is_null());
     let response: &CommandResponse = unsafe { &*response_ptr };
     assert!(!response.string_value.is_null());
-    let mut map = RESULTS_MAP
-        .lock()
-        .expect("Failed to aquire the results' lock");
     // Create a byte slice from the raw pointer
     let bytes = unsafe {
         std::slice::from_raw_parts(
@@ -39,31 +57,19 @@ extern "C" fn string_success_callback(index: usize, response_ptr: *const Command
         )
     };
     // Convert bytes to a Rust string, handling UTF-8 safely
-    let str_response = std::str::from_utf8(bytes)
+    std::str::from_utf8(bytes)
         .map(|s| s.to_owned())
-        .expect("Failed to parse string value");
-    map.insert(index, Ok(str_response));
-
-    SUCCESS_COUNTER.fetch_add(1, Ordering::SeqCst);
+        .expect("Failed to parse string value")
 }
 
-/// Failure callback function
-extern "C" fn failure_callback(
-    index: usize,
-    err_msg_ptr: *const c_char,
-    error_type: RequestErrorType,
-) {
+fn parse_error_msg(err_msg_ptr: *const c_char) -> String {
     assert!(!err_msg_ptr.is_null());
     let err_msg: &'static str = unsafe {
         CStr::from_ptr(err_msg_ptr)
             .to_str()
             .expect("Failed to parse err_msg")
     };
-    let mut map = RESULTS_MAP
-        .lock()
-        .expect("Failed to aquire the results' lock");
-    map.insert(index, Err((err_msg, error_type)));
-    FAILURE_COUNTER.fetch_add(1, Ordering::SeqCst);
+    err_msg.to_string()
 }
 
 // TODO: Move RedisServer implementation from glide-core tests to a reusable library and replace this Server implementation.
@@ -138,7 +144,7 @@ fn execute_command(
     command_bytes: &[u8],
     arg_count: u64,
     command_type: RequestType,
-) {
+) -> Option<Box<CommandResult>> {
     let command_len = command_bytes.len();
     // Convert command arguments into a vector of pointers
     let command_vec = [command_bytes.as_ptr()];
@@ -150,8 +156,7 @@ fn execute_command(
     let empty_data: Vec<u8> = vec![];
     let route_bytes: *const u8 = empty_data.as_ptr();
     let route_len = 0;
-
-    unsafe {
+    let command_res_ptr = unsafe {
         command(
             client_ptr,
             index,
@@ -161,17 +166,27 @@ fn execute_command(
             args_len_ptr,
             route_bytes,
             route_len,
-        );
+        )
+    };
+    if command_res_ptr.is_null() {
+        // Async client is being used, let the async callback to be called
+        let rt = Runtime::new().unwrap();
+        rt.block_on(async {
+            sleep(Duration::from_millis(1)).await;
+        });
+        None
+    } else {
+        // Sync client, command result returns immediately
+        unsafe { Some(Box::from_raw(command_res_ptr)) }
     }
-    let rt = Runtime::new().unwrap();
-    rt.block_on(async {
-        // Let the async callback to be called
-        sleep(Duration::from_millis(1)).await;
-    });
 }
 
-fn get_response(index: usize) -> String {
-    let map = RESULTS_MAP.lock().unwrap();
+fn get_sync_response(cmd_resp: *mut CommandResponse) -> String {
+    parse_string_res(cmd_resp)
+}
+
+fn get_async_response(index: usize) -> String {
+    let map = ASYNC_RESULTS_MAP.lock().unwrap();
     let result = map
         .get(&index)
         .expect("Couldn't find the relevant idx in the map");
@@ -179,8 +194,17 @@ fn get_response(index: usize) -> String {
     result.as_ref().expect("Can't unwrap result").to_string()
 }
 
-fn get_error(index: usize) -> (String, RequestErrorType) {
-    let map = RESULTS_MAP.lock().unwrap();
+fn get_sync_error(command_error: *mut CommandError) -> (String, RequestErrorType) {
+    assert!(!command_error.is_null());
+    let command_err: Box<CommandError> = unsafe { Box::from_raw(command_error) };
+    (
+        parse_error_msg(command_err.command_error_message),
+        command_err.command_error_type,
+    )
+}
+
+fn get_async_error(index: usize) -> (String, RequestErrorType) {
+    let map = ASYNC_RESULTS_MAP.lock().unwrap();
     let result = map
         .get(&index)
         .expect("Couldn't find the relevant idx in the map");
@@ -189,19 +213,24 @@ fn get_error(index: usize) -> (String, RequestErrorType) {
     (err_msg.to_string(), err_type.clone())
 }
 
-#[test]
-fn test_command_execution() {
+#[rstest]
+#[serial_test::serial]
+fn test_ffi_client_command_execution(#[values(false, true)] async_client: bool) {
     let server = Server::new();
     let connection_request_bytes = create_connection_request(server.port);
     let connection_request_len = connection_request_bytes.len();
     let connection_request_ptr = connection_request_bytes.as_ptr();
-    unsafe {
-        let response_ptr = create_client(
-            connection_request_ptr,
-            connection_request_len,
-            string_success_callback,
+    let client_type = Box::into_raw(Box::new(if async_client {
+        ClientType::AsyncClient {
+            success_callback: string_success_callback,
             failure_callback,
-        );
+        }
+    } else {
+        ClientType::SyncClient
+    }));
+    unsafe {
+        let response_ptr =
+            create_client(connection_request_ptr, connection_request_len, client_type);
 
         assert!(!response_ptr.is_null(), "Failed to create client");
         let response = &*response_ptr;
@@ -214,29 +243,39 @@ fn test_command_execution() {
         // Good command: PING IS_WORKING
         let good_cmd_idx = 0;
         let ping_value = b"IS_WORKING";
-        execute_command(
+        let good_res = execute_command(
             client_ptr,
             good_cmd_idx,
             ping_value,
             1_u64,
             RequestType::Ping,
         );
-        assert_eq!(SUCCESS_COUNTER.load(Ordering::SeqCst), 1);
-        assert_eq!(
-            get_response(good_cmd_idx),
-            String::from_utf8_lossy(ping_value)
-        );
+        let ping_res = if async_client {
+            assert!(good_res.is_none()); // result should be returned through callback
+            assert_eq!(ASYNC_SUCCESS_COUNTER.load(Ordering::SeqCst), 1);
+            get_async_response(good_cmd_idx)
+        } else {
+            assert!(good_res.is_some());
+            get_sync_response(good_res.unwrap().response)
+        };
+        assert_eq!(ping_res, String::from_utf8_lossy(ping_value));
         // Bad command: Non existing command args, the server should return with an error: SADD NOTAREALCMD
         let bad_cmd_idx = 1;
-        execute_command(
+        let bad_res = execute_command(
             client_ptr,
             bad_cmd_idx,
             b"NOTAREALCMD",
             1_u64,
             RequestType::SAdd,
         );
-        assert_eq!(FAILURE_COUNTER.load(Ordering::SeqCst), 1);
-        let (err_msg, err_type) = get_error(bad_cmd_idx);
+        let (err_msg, err_type) = if async_client {
+            assert!(bad_res.is_none()); // result should be returned through callback
+            assert_eq!(ASYNC_FAILURE_COUNTER.load(Ordering::SeqCst), 1);
+            get_async_error(bad_cmd_idx)
+        } else {
+            assert!(bad_res.is_some());
+            get_sync_error(bad_res.unwrap().command_error)
+        };
         assert!(err_msg.contains("wrong number of arguments for 'sadd' command"));
         assert_eq!(err_type, RequestErrorType::Unspecified);
         free_connection_response(response_ptr as *mut ConnectionResponse);